Fix network.rs in order to match new libogc modifications

+++ src/network.rs
@@ -183,7 +183,7 @@ pub struct SocketAddress {
     // The address family.
     pub family: ProtocolFamily,
     // The address data.
-    pub data: [u8; 14],
+    pub data: [i8; 14],
 }
 
 /// Convert ``SocketAddress`` into a ``ogc_sys::sockaddr``.
@@ -208,9 +208,9 @@ pub struct HostInformation {
     /// A NULL-terminated array of alternate names.
     pub aliases: Vec<String>,
     /// The type of address being returned.
-    pub address_type: i32,
+    pub address_type: u16,
     /// The length, in bytes, of each address.
-    pub length: i32,
+    pub length: u16,
     /// A NULL-terminated list of addresses for the host.
     pub address_list: Vec<String>,
 }
